Move waitUntilNextFramePainted from common into lib utils
