+++ frontend/src/metabase/documents/hooks/use-print-context-value.ts
@@ -1,9 +1,9 @@
 import { useWindowEvent } from "@mantine/hooks";
 import { useCallback, useMemo, useState } from "react";
 
-import { waitUntilNextFramePainted } from "metabase/common/utils/wait-until-next-frame-paints";
 import type { PrintContextValue } from "metabase/documents/contexts/PrintContext";
 import { delay } from "metabase/utils/promise";
+import { waitUntilNextFramePainted } from "metabase/utils/wait-until-next-frame-paints";
 
 const READINESS_TIMEOUT_MS = 15_000;
 const POLL_INTERVAL_MS = 100;
