+++ frontend/src/metabase/utils/wait-until-next-frame-paints.ts
